Use u64 to track file offsets instead of usize (#2573)
* use u64 internally to track file offsets

This replaces usize in a lot of places. Besides
being a technically breaking change, this should
not impact functionality.
Solves a TODO left long ago in the entry code
of parse.

* inline LogicalOffset as per review comment

* self-review to simplify logic slightly

* convert wasmprinter and wasm-encoder

* convert wasm-metadata, wasm-mutate, wast (tests)

* fix msrv issue

* convert wasm-tools bins and fuzz

* fix some lints and self-review

* fix clippy lints and adjust test on error message

* remove and inline MemOffset structure by review

* remove InMemData from two uses

* move range conversion into an OffsetCoverter struct

this is now part of the return from the parse functions, so it's
impossible to miss if you need access to it

* remove OffsetConverter from parse_all

Almost all of the time, parse_all is called at offset 0.
In these cases, the conversion is simplified and not needed.

* remove references to OffsetConverter::from_start(0)

Use 'simply' as-casts instead. In these cases, we simply assert
that our casting logic is sound and trust in the parser to not
emit offsets that silently truncate

* oops missed one due to not building wit-dylib locally

* convert wasm-tools to as-casts

* remove ConvertOffset

what's a few as casts among friends. certainly too simple
of an API to leave in, right?

* minify the diff by reverting some unrelated changes

* two more stylistic changes for diff optimization

* fix clippy warnings

`u32::max_value()` et al got deprecated last week as part of
rust-lang/rust#160517 in rust-lang/rust#146882

* do not silently truncate data input

the parser will return an Error, the binary reader will panic

+++ crates/wasm-metadata/src/payload.rs
@@ -40,10 +40,10 @@ impl Payload {
                     if output.is_empty() {
                         match encoding {
                             wasmparser::Encoding::Module => {
-                                output.push(Self::empty_module(0..input.len()))
+                                output.push(Self::empty_module(0..input.len() as u64))
                             }
                             wasmparser::Encoding::Component => {
-                                output.push(Self::empty_component(0..input.len()))
+                                output.push(Self::empty_component(0..input.len() as u64))
                             }
                         }
                     }
@@ -191,7 +191,7 @@ impl Payload {
         }
     }
 
-    fn empty_component(range: Range<usize>) -> Self {
+    fn empty_component(range: Range<u64>) -> Self {
         let mut this = Self::Component {
             metadata: Metadata::default(),
             children: vec![],
@@ -200,7 +200,7 @@ impl Payload {
         this
     }
 
-    fn empty_module(range: Range<usize>) -> Self {
+    fn empty_module(range: Range<u64>) -> Self {
         let mut this = Self::Module(Metadata::default());
         this.metadata_mut().range = range;
         this
