+++ crates/wasmparser/src/error.rs
@@ -16,7 +16,7 @@ pub struct Error {
 pub(crate) struct ErrorInner {
     message: String,
     kind: ErrorKind,
-    offset: usize,
+    offset: u64,
     needed_hint: Option<usize>,
 }
 
@@ -44,7 +44,7 @@ impl fmt::Display for Error {
 
 impl Error {
     #[cold]
-    pub(crate) fn _new(kind: ErrorKind, message: String, offset: usize) -> Self {
+    pub(crate) fn _new(kind: ErrorKind, message: String, offset: u64) -> Self {
         Error {
             inner: Box::new(ErrorInner {
                 kind,
@@ -56,12 +56,12 @@ impl Error {
     }
 
     #[cold]
-    pub(crate) fn new(message: impl Into<String>, offset: usize) -> Self {
+    pub(crate) fn new(message: impl Into<String>, offset: u64) -> Self {
         Self::_new(ErrorKind::Uncategorized, message.into(), offset)
     }
 
     #[cold]
-    pub(crate) fn invalid_heap_type(msg: &'static str, offset: usize) -> Self {
+    pub(crate) fn invalid_heap_type(msg: &'static str, offset: u64) -> Self {
         Self::_new(ErrorKind::InvalidHeapType, msg.into(), offset)
     }
 
@@ -69,18 +69,18 @@ impl Error {
     pub(crate) fn wasm_feature(
         feature: crate::WasmFeatures,
         msg: impl fmt::Display,
-        offset: usize,
+        offset: u64,
     ) -> Self {
         Self::_new(ErrorKind::WasmFeature(feature), msg.to_string(), offset)
     }
 
     #[cold]
-    pub(crate) fn fmt(args: fmt::Arguments<'_>, offset: usize) -> Self {
+    pub(crate) fn fmt(args: fmt::Arguments<'_>, offset: u64) -> Self {
         Error::new(args.to_string(), offset)
     }
 
     #[cold]
-    pub(crate) fn eof(offset: usize, needed_hint: usize) -> Self {
+    pub(crate) fn eof(offset: u64, needed_hint: usize) -> Self {
         let mut err = Error::new("unexpected end-of-file", offset);
         err.inner.needed_hint = Some(needed_hint);
         err
@@ -96,7 +96,7 @@ impl Error {
     }
 
     /// Get the offset within the Wasm binary where the error occurred.
-    pub fn offset(&self) -> usize {
+    pub fn offset(&self) -> u64 {
         self.inner.offset
     }
